NetworkStateHandler: Add portal state logging
This adds a << operator for PortalState and some additional
portal state logging to help with debugging.

BUG=b:207069182
TEST=manual log inspection

Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3811690
Commit-Queue: Steven Bennetts <stevenjb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1034099}

+++ chromeos/ash/components/network/network_state_handler_unittest.cc
@@ -144,7 +144,7 @@ class TestObserver final : public NetworkStateHandlerObserver {
                           NetworkState::PortalState portal_state) override {
     default_network_portal_state_ = portal_state;
     ++portal_state_change_count_;
-    VLOG(1) << "PortalStateChanged: " << static_cast<int>(portal_state);
+    VLOG(1) << "PortalStateChanged: " << portal_state;
   }
 
   void NetworkConnectionStateChanged(const NetworkState* network) override {
